Implementação da orientação a objetos ao MultiData Toda a funcionalidade do Multidata quebrada em multiplos arquivos diferentes com objetos diferentes
MultiData agora é uma classe de interface
